@@ -1,13 +1,13 @@
 source "https://rubygems.org"
 
-gem "rails", "~> 5.1.0"
+gem "rails", "~> 5.2.0"
 
-gem "coffee-rails", "~> 4.2" # Rails 5.1 default
+gem "coffee-rails", "~> 4.2" # Rails 5.2 default
 gem "jquery-rails"
 gem "jquery-ui-rails"
-gem "uglifier", ">= 1.3.0" # Rails 5.1 default
+gem "uglifier", ">= 1.3.0" # Rails 5.2 default
 
-gem "sass-rails", "~> 5.0" # Rails 5.1 default
+gem "sass-rails", "~> 5.0" # Rails 5.2 default
 
 source "https://rails-assets.org" do
   gem "rails-assets-js-cookie"
@@ -55,7 +55,7 @@ end
 
 gem "pry", :group => [:development, :test]
 
-gem "jbuilder", "~> 2.5" # Rails 5.1 default
+gem "jbuilder", "~> 2.5" # Rails 5.2 default
 
 # Must be last.
 gem "rack-mini-profiler", :group => :development
